requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.7
curlify==3.0.0
dotenv==0.9.9
facebook_business==25.0.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-ads==30.1.0
google-api-core==2.30.3
google-auth==2.49.2
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
greenlet==3.3.2
grpcio==1.80.0
grpcio-status==1.80.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
multidict==6.7.1
narwhals==2.17.0
numpy==2.4.2
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pillow==12.1.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycountry==26.2.16
pycparser==3.0
pydeck==0.9.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.48
streamlit==1.55.0
streamlit-echarts==0.6.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
yarl==1.23.0
